cleanup: break shapes into files
- rename Shapes.drawAll -> Shapes.draw
- make old Shapes.draw a private method in shapes/draw.js
- make old Shapes.convertPath a private method in shapes/draw.js
- rm Shapes.add (wasn't used anywhere)
